Hair Restyle Deed
You used to get this deed with the newbie ticket from t2a when you and another ticket was able to be combined together. When you started a new char on any server.

This deed will work for male and female chars no other race. as it was before elves and gargoyles.