Add Door Command
This command will bring up a gump for easy selection of doors,

when you click the Arrow on top of the door it will bring up a target cursor you just need to point to the location and it will repeat the command until you cancel the target.

This command also remembers the last page you was on. so when  you do do 'adddoor again it will bring you right back tot he page you left off.