Solen Ant Hole Add-On
This was introduced in LBR.
The script will allow the user to double click the middle black hole and teleport to the solen hive.

This will also spawn 2 to 5 ants when the hole is walked on.

Keeps track of the deaths of each ant so it cant over spawn the area.