Added Map RuleSet
This can make it so players can not attack steal or cast on other players why in trammel ,malas and  tokuno or ilsh.

The gate will attach the script to the player when they go though gate to those facets. if the player goes back to the felucca the script is removed.