Update mapitem.js
this will no longer draw the map out with blank maps since the map is now crated using the crafting skill. I left the function in place just incase a user would like to maybe have blank maps drawn from map item and not a craft menu. but it is disabled.