recipe system
This simple custom recipe system. with example recipe.

When you double click scorll to read it will make a save text file just for that char and all the recipes he learns. this recipes will need to be added to each button press for the system to read if not it will always be 0, example is added using plate chest.