more refactoring
More Details on Each function
ToolTip Update to show full status of the plant when mouse over.
applypotion update
Update to PlantImage and PlantBowl
Stage 14 dirt added to all places
Cleanup of tags in some areas for easier managing
Fixed if the plant is not locked down it will not Grow.